@@ -24,86 +24,86 @@
     "jquery": "3.4.1",
     "keymaster": "1.6.2",
     "lz-string": "1.4.4",
-    "markdown-it": "9.0.1",
+    "markdown-it": "10.0.0",
     "node-noop": "1.0.0",
     "po2json": "0.4.5",
-    "react": "16.8.6",
-    "react-dom": "16.8.6",
-    "react-redux": "7.1.0",
-    "react-router": "5.0.1",
-    "react-router-dom": "5.0.1",
-    "redux": "4.0.4",
+    "react": "16.12.0",
+    "react-dom": "16.12.0",
+    "react-redux": "7.1.3",
+    "react-router": "5.1.2",
+    "react-router-dom": "5.1.2",
+    "redux": "4.0.5",
     "redux-devtools-extension": "2.13.8",
     "redux-thunk": "2.3.0",
     "reflect-metadata": "0.1.13",
     "tslib": "1.10.0",
-    "typesafe-actions": "4.4.2"
+    "typesafe-actions": "5.1.0"
   },
   "devDependencies": {
-    "@types/chai": "4.1.7",
-    "@types/chai-as-promised": "7.1.0",
+    "@types/chai": "4.2.7",
+    "@types/chai-as-promised": "7.1.2",
     "@types/classnames": "2.2.9",
     "@types/clipboard": "2.0.1",
-    "@types/codemirror": "0.0.76",
-    "@types/color": "3.0.0",
+    "@types/codemirror": "0.0.83",
+    "@types/color": "3.0.1",
     "@types/core-js": "2.5.2",
-    "@types/enzyme": "3.10.3",
+    "@types/enzyme": "3.10.4",
     "@types/enzyme-adapter-react-16": "1.0.5",
     "@types/file-saver": "2.0.1",
-    "@types/history": "4.7.2",
+    "@types/history": "4.7.4",
     "@types/isomorphic-fetch": "0.0.35",
-    "@types/jquery": "3.3.30",
+    "@types/jquery": "3.3.31",
     "@types/jsdom": "12.2.4",
     "@types/keymaster": "1.6.28",
     "@types/lz-string": "1.3.33",
-    "@types/markdown-it": "0.0.8",
+    "@types/markdown-it": "0.0.9",
     "@types/mocha": "5.2.7",
-    "@types/node": "12.6.8",
-    "@types/react": "16.8.23",
-    "@types/react-dom": "16.8.5",
-    "@types/react-redux": "7.1.1",
-    "@types/react-router": "5.0.3",
-    "@types/react-router-dom": "4.3.4",
-    "@types/react-test-renderer": "16.8.3",
+    "@types/node": "13.5.0",
+    "@types/react": "16.9.19",
+    "@types/react-dom": "16.9.5",
+    "@types/react-redux": "7.1.7",
+    "@types/react-router": "5.1.4",
+    "@types/react-router-dom": "5.1.3",
+    "@types/react-test-renderer": "16.9.2",
     "@types/redux-mock-store": "1.0.1",
-    "@types/sinon": "7.0.13",
+    "@types/sinon": "7.5.1",
     "chai": "4.2.0",
     "chai-as-promised": "7.1.1",
     "clean-webpack-plugin": "3.0.0",
-    "copy-webpack-plugin": "5.0.4",
-    "css-loader": "3.1.0",
-    "enzyme": "3.10.0",
-    "enzyme-adapter-react-16": "1.14.0",
-    "file-loader": "4.1.0",
-    "fork-ts-checker-webpack-plugin": "1.4.3",
-    "git-revision-webpack-plugin": "3.0.3",
+    "copy-webpack-plugin": "5.1.1",
+    "css-loader": "3.4.2",
+    "enzyme": "3.11.0",
+    "enzyme-adapter-react-16": "1.15.2",
+    "file-loader": "5.0.2",
+    "fork-ts-checker-webpack-plugin": "4.0.2",
+    "git-revision-webpack-plugin": "3.0.4",
     "html-webpack-plugin": "3.2.0",
-    "husky": "3.0.2",
+    "husky": "4.2.1",
     "ignore-styles": "5.0.1",
-    "jsdom": "15.1.1",
+    "jsdom": "16.0.1",
     "json-loader": "0.5.7",
-    "less": "3.9.0",
+    "less": "3.10.3",
     "less-loader": "5.0.0",
-    "mini-css-extract-plugin": "0.8.0",
-    "mocha": "6.2.0",
+    "mini-css-extract-plugin": "0.9.0",
+    "mocha": "7.0.0",
     "pofile": "1.0.11",
-    "prettier": "1.18.2",
-    "pretty-quick": "1.11.1",
-    "raw-loader": "3.1.0",
-    "redux-mock-store": "1.5.3",
-    "sinon": "7.3.2",
-    "style-loader": "0.23.1",
-    "ts-loader": "6.0.4",
-    "ts-node": "8.3.0",
-    "tsconfig-paths": "3.8.0",
-    "tslint": "5.18.0",
+    "prettier": "1.19.1",
+    "pretty-quick": "2.0.1",
+    "raw-loader": "4.0.0",
+    "redux-mock-store": "1.5.4",
+    "sinon": "8.1.1",
+    "style-loader": "1.1.3",
+    "ts-loader": "6.2.1",
+    "ts-node": "8.6.2",
+    "tsconfig-paths": "3.9.0",
+    "tslint": "6.0.0",
     "tslint-sonarts": "1.9.0",
-    "typescript": "3.5.3",
-    "url-loader": "2.1.0",
-    "webpack": "4.38.0",
-    "webpack-bundle-analyzer": "3.4.1",
-    "webpack-cli": "3.3.6",
-    "webpack-dev-server": "3.7.2"
+    "typescript": "3.7.5",
+    "url-loader": "3.0.0",
+    "webpack": "4.41.5",
+    "webpack-bundle-analyzer": "3.6.0",
+    "webpack-cli": "3.3.10",
+    "webpack-dev-server": "3.10.1"
   },
   "husky": {
     "hooks": {
